+++ src/main/java/org/opensearch/knn/index/store/IndexOutputWithBuffer.java
@@ -8,7 +8,12 @@ package org.opensearch.knn.index.store;
 import org.apache.lucene.store.IndexOutput;
 
 import java.io.IOException;
+import java.io.InputStream;
 
+/**
+ * Wrapper around {@link IndexOutput} to perform writes in a buffered manner. This class is created per flush/merge, and may be used twice if
+ * {@link org.opensearch.knn.index.codec.nativeindex.remote.RemoteIndexBuildStrategy} needs to fall back to a different build strategy.
+ */
 public class IndexOutputWithBuffer {
     // Underlying `IndexOutput` obtained from Lucene's Directory.
     private IndexOutput indexOutput;
@@ -16,10 +21,12 @@ public class IndexOutputWithBuffer {
     // Allocating 64KB here since it show better performance in NMSLIB with the size. (We had slightly improvement in FAISS than having 4KB)
     // NMSLIB writes an adjacent list size first, then followed by serializing the list. Since we usually have more adjacent lists, having
     // 64KB to accumulate bytes as possible to reduce the times of calling `writeBytes`.
-    private byte[] buffer = new byte[64 * 1024];
+    private static final int CHUNK_SIZE = 64 * 1024;
+    private final byte[] buffer;
 
     public IndexOutputWithBuffer(IndexOutput indexOutput) {
         this.indexOutput = indexOutput;
+        this.buffer = new byte[CHUNK_SIZE];
     }
 
     // This method will be called in JNI layer which precisely knows
@@ -30,6 +37,43 @@ public class IndexOutputWithBuffer {
             indexOutput.writeBytes(buffer, 0, length);
         } catch (IOException e) {
             throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Writes to the {@link IndexOutput} by buffering bytes into the existing buffer in this class.
+     *
+     * @param inputStream       The stream from which we are reading bytes to write
+     * @throws IOException
+     * @see IndexOutputWithBuffer#writeFromStreamWithBuffer(InputStream, byte[])
+     */
+    public void writeFromStreamWithBuffer(InputStream inputStream) throws IOException {
+        writeFromStreamWithBuffer(inputStream, this.buffer);
+    }
+
+    /**
+     * Writes to the {@link IndexOutput} by buffering bytes with @param outputBuffer. This method allows
+     * {@link org.opensearch.knn.index.codec.nativeindex.remote.RemoteIndexBuildStrategy} to provide a separate, larger buffer as that buffer is for buffering
+     * bytes downloaded from the repository, so it may be more performant to use a larger buffer.
+     * We do not change the size of the existing buffer in case a fallback to the existing build strategy is needed.
+     * TODO: Tune the size of the buffer used by RemoteIndexBuildStrategy based on benchmarking
+     *
+     * @param inputStream       The stream from which we are reading bytes to write
+     * @param outputBuffer      The buffer used to buffer bytes
+     * @throws IOException
+     * @see IndexOutputWithBuffer#writeFromStreamWithBuffer(InputStream)
+     */
+    private void writeFromStreamWithBuffer(InputStream inputStream, byte[] outputBuffer) throws IOException {
+        int bytesRead = 0;
+        // InputStream uses -1 indicates there are no more bytes to be read
+        while (bytesRead != -1) {
+            // Try to read CHUNK_SIZE into the buffer. The actual amount read may be less.
+            bytesRead = inputStream.read(outputBuffer, 0, CHUNK_SIZE);
+            assert bytesRead <= CHUNK_SIZE;
+            // However many bytes we read, write it to the IndexOutput if != -1
+            if (bytesRead != -1) {
+                indexOutput.writeBytes(outputBuffer, 0, bytesRead);
+            }
         }
     }